Fix the app init that occured due to side effects when running pytest
